requirements/base.txt
# Django Core
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0,<3.15.0
django-cors-headers>=4.3.0,<4.4.0
django-filter>=23.3,<23.4

# Database
psycopg2-binary>=2.9.9,<2.10.0
django-mptt>=0.14.0,<0.15.0

# Authentication & Authorization
djangorestframework-simplejwt>=5.3.0,<5.4.0
django-allauth>=0.57.0,<0.58.0

# API Documentation
drf-spectacular>=0.26.0,<0.27.0

# Task Queue
celery>=5.3.0,<5.4.0
redis>=5.0.0,<5.1.0

# Search
elasticsearch>=8.10.0,<8.11.0

# Real-time
channels>=4.0.0,<4.1.0
channels-redis>=4.1.0,<4.2.0

# Utilities
python-dotenv>=1.0.0,<1.1.0
django-environ>=0.11.0,<0.12.0
Pillow>=10.0.0,<10.1.0
pycountry>=22.3.5,<22.4.0

# Flags
django-flags>=5.0.0,<5.1.0

# AWS
boto3>=1.34.0,<1.35.0
